+++ src/Aquifer.sol
@@ -17,7 +17,7 @@ import {LibClone} from "src/libraries/LibClone.sol";
  * @dev Aquifer deploys Wells by cloning a pre-deployed Well implementation.
  */
 contract Aquifer is IAquifer, ReentrancyGuard {
-    using SafeCast for uint;
+    using SafeCast for uint256;
     using LibClone for address;
 
     // A mapping of Well address to the Well implementation addresses
@@ -27,7 +27,7 @@ contract Aquifer is IAquifer, ReentrancyGuard {
     constructor() ReentrancyGuard() {}
 
     /**
-     * @dev see {IAquifer.boreWell}
+     * @dev
      * Use `salt == 0` to deploy a new Well with `create`
      * Use `salt > 0` to deploy a new Well with `create2`
      */
@@ -83,9 +83,6 @@ contract Aquifer is IAquifer, ReentrancyGuard {
         );
     }
 
-    /**
-     * @dev see {IAquifer.wellImplementation}
-     */
     function wellImplementation(address well) external view returns (address implementation) {
         return wellImplementations[well];
     }
